--- OpenMarket/OpenMarket/Network/DeleteItemData.swift
@@ -1,12 +1,0 @@
-//
-//  DeleteItemData.swift
-//  OpenMarket
-//
-//  Created by tae hoon park on 2021/08/16.
-//
-
-import Foundation
-
-struct DeleteItemData: Encodable {
-    let password: String
-}
