+++ OpenMarket/OpenMarket/Protocol/NetworkProtocol.swift
@@ -11,8 +11,6 @@ protocol URLSessionProtocol {
     func dataTask(with request: URLRequest, completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) -> URLSessionDataTask
 }
 
-extension URLSession: URLSessionProtocol {
-    
-}
+extension URLSession: URLSessionProtocol {}
